Fix SCL root password handling in deploy script

+++ deploy-to-production.ps1
@@ -146,7 +146,10 @@ if ($SyncSclDb) {
 		throw "Failed to detect MYSQL_ROOT_PASSWORD from scli-mysql-prod"
 	}
 
-	$pwdEscaped = $rootPwd.Replace("'", "'\"'\"'")
+	if ($rootPwd.Contains("'")) {
+		throw "MYSQL_ROOT_PASSWORD contains a single quote, which this script currently does not support"
+	}
+	$pwdEscaped = $rootPwd
 	$sclRemoteSql = "docker exec scli-mysql-prod mysql -uroot -p'$pwdEscaped' -e \"DROP DATABASE IF EXISTS $SclDatabase;\"; " +
 					"docker exec scli-mysql-prod mysql -uroot -p'$pwdEscaped' -e \"CREATE DATABASE $SclDatabase CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci;\"; " +
 					"docker exec -i scli-mysql-prod mysql -uroot -p'$pwdEscaped' $SclDatabase < /tmp/scl_sync.sql"
